requirements.txt
--extra-index-url https://k2-fsa.github.io/sherpa/onnx/cuda.html

# 语音识别和音频处理
librosa==0.11.0
numba==0.61.2
numpy==2.2.6
sounddevice==0.5.2
soundfile==0.13.1
sherpa_onnx
pypinyin

# 全局快捷键和文本输入
keyboard>=0.13.5
pyautogui>=0.9.54
pyperclip>=1.8.2

# 配置管理
pyyaml>=6.0.1

# 系统监控
psutil>=5.9.0

# 可选依赖 - 平台特定
pywin32>=306; sys_platform == "win32"
pyobjc>=9.0; sys_platform == "darwin"

# 已有依赖（可能用于其他功能）
pydantic
pydantic-settings
fastapi
uvicorn
uvicorn[standard]
jinja2
scipy
ffmpeg-python
python-multipart 
loguru
